Render the node-graph on the template browser cards
The template browser (/admin/aabenforms/workflow-templates) showed broken
bpmn-js thumbnails stuck on "Loading preview...". Replace them with the shared
node-graph: every template X maps 1:1 to ECA flow X_flow, so each card now
renders that flow's real structure (start -> gated steps -> deny branch) from
EcaGraphBuilder.

- Extract the SVG rendering out of FlowGraphController into a shared
  FlowGraphRenderer service (render + renderForFlow), used by both the per-flow
  page and the cards so they draw identically.
- TemplateBrowserController renders each template's flow graph into its card and
  points "View Full Preview" at the full flow diagram page; falls back to the
  legacy BPMN preview when a template has no flow yet.
- Card sizes the diagram to the thumbnail height and scrolls horizontally for
  wider flows rather than shrinking to dots.
